Refactor tests to avoid safari webdriver pairing issue
