+++ daily_quote/tests.py
@@ -1,5 +1,9 @@
+from os import environ
+
 from django.contrib.auth.models import User
 from django.test import TestCase
+from selenium import webdriver
+from selenium.webdriver import DesiredCapabilities
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support import expected_conditions
 from selenium.webdriver.support.wait import WebDriverWait
@@ -61,6 +65,33 @@ class QuoteRecommendationTests(TestCase):
 
 
 class FunctionalTests(FunctionalTestCase):
+    def setUp(self):
+        User.objects.create_user('alice', 'alice@gmail.com', 'pass1234')
+
+        if "TRAVIS" in environ:
+            username = environ["SAUCE_USERNAME"]
+            access_key = environ["SAUCE_ACCESS_KEY"]
+            # Create a desired capabilities object as a starting point.
+            capabilities = DesiredCapabilities.FIREFOX.copy()
+            capabilities['platform'] = "WINDOWS"
+            capabilities["tunnel-identifier"] = environ["TRAVIS_JOB_NUMBER"]
+            capabilities["build"] = environ["TRAVIS_BUILD_NUMBER"]
+            capabilities["tags"] = [environ["TRAVIS_PYTHON_VERSION"], "CI"]
+            hub_url = "%s:%s@localhost:4445" % (username, access_key)
+            self.selenium = webdriver.Remote(desired_capabilities=capabilities,
+                                             command_executor="http://%s/wd/hub" % hub_url)
+        else:
+            self.selenium = webdriver.Safari()
+            super(FunctionalTestCase, self).setUp()
+
+    def tearDown(self):
+        # QuoteRank.objects.all().delete()
+        # Profile.objects.all().delete()
+        # User.objects.all().delete()
+
+        self.selenium.quit()
+        super(FunctionalTestCase, self).tearDown()
+
     def test_like_button_profile(self):
         selenium = self.selenium
         selenium.maximize_window()
