+++ quote_me/tests.py
@@ -77,33 +77,6 @@ class FunctionalTestCase(LiveServerTestCase):
     login_url = 'http://127.0.0.1:8000/login/'
     logout_url = 'http://127.0.0.1:8000/logout/'
 
-    def setUp(self):
-        User.objects.create_user('alice', 'alice@gmail.com', 'pass1234')
-
-        if "TRAVIS" in environ:
-            username = environ["SAUCE_USERNAME"]
-            access_key = environ["SAUCE_ACCESS_KEY"]
-            # Create a desired capabilities object as a starting point.
-            capabilities = DesiredCapabilities.FIREFOX.copy()
-            capabilities['platform'] = "WINDOWS"
-            capabilities["tunnel-identifier"] = environ["TRAVIS_JOB_NUMBER"]
-            capabilities["build"] = environ["TRAVIS_BUILD_NUMBER"]
-            capabilities["tags"] = [environ["TRAVIS_PYTHON_VERSION"], "CI"]
-            hub_url = "%s:%s@localhost:4445" % (username, access_key)
-            self.selenium = webdriver.Remote(desired_capabilities=capabilities,
-                                             command_executor="http://%s/wd/hub" % hub_url)
-        else:
-            self.selenium = webdriver.Safari()
-            super(FunctionalTestCase, self).setUp()
-
-    def tearDown(self):
-        # QuoteRank.objects.all().delete()
-        # Profile.objects.all().delete()
-        # User.objects.all().delete()
-
-        self.selenium.quit()
-        super(FunctionalTestCase, self).tearDown()
-
     def signup(self):
         selenium = self.selenium
         selenium.get(self.signup_url)
